requirements.txt
apache-airflow==2.9.2
mysql-connector
mysql-connector-python
pandas
psycopg2-binary==2.9.9
requests==2.32.3
apache-airflow-providers-common-sql
apache-airflow-providers-postgres
apache-airflow-providers-mysql
apache-airflow-providers-google
openpyxl